Add Scripts to Unlock Directory for Twitch App
Git locks the directory in the eyes of the Twitch App
(for some unknown reason). In order to fix this so that
the "builder" profile can be updated to the latest
version of the modpack it is necessary to remove the git
tracking information, update the profile, and the re-add
the git tracking information. These scripts help automate
this process.

Note: NEEDS to be done after any changes made are uploaded
to the remote or changes will be PERMANENTLY lost.